+++ docs/Whats-New.md
@@ -93,6 +93,8 @@ Vanilla fixes:
 - Fix a bug where the "Building exists" event would fire when you queued a building on the sidebar (by ZivDero)
 - Fix a bug where using the "Destroy Tag" trigger action could lead to trying to free invalid memory (by ZivDero)
 - Fix a bug where the game could freeze in the score screen in `Clip_Line` when running on Windows 11 24H2 (by tomsons26, Rampastring)
+- Fix a bug where carryalls would unload infantry like they unload vehicles (by ZivDero)
+- Fix a bug where carryalls would draw their infantry passenger's shadow (by ZivDero)
 
 New:
 - Implement `BarGate` for buildings (by ZivDero)